+++ data-models/ModelDefinitionsSet.ts
@@ -197,13 +197,13 @@ module ModelDefinitionsSet {
     export function cloneDtoPropertyTemplate(prop: DtoPropertyTemplate & PropertyConversionTemplate, cloneDeep: <T1>(obj: T1) => T1 = Objects.cloneDeep): DtoPropertyTemplate & PropertyConversionTemplate {
         return {
             autoGenerate: prop.autoGenerate,
-            defaultValue: prop.defaultValue != null ? cloneDeep(prop.defaultValue) : null,
+            defaultValue: prop.defaultValue != null ? cloneDeep(prop.defaultValue) : prop.defaultValue,
             primaryKey: prop.primaryKey,
             readOnly: prop.readOnly,
             required: prop.required,
             server: prop.server == null ? null : {
                 autoGenerate: prop.server.autoGenerate,
-                defaultValue: prop.server.defaultValue != null ? cloneDeep(prop.server.defaultValue) : null,
+                defaultValue: prop.server.defaultValue != null ? cloneDeep(prop.server.defaultValue) : prop.server.defaultValue,
                 name: prop.server.name,
                 primaryKey: prop.server.primaryKey,
                 readOnly: prop.server.readOnly,