+++ db-collections/in-mem-collections.d.ts
@@ -1,4 +1,5 @@
 ﻿/// <reference path="../../definitions/lib/lokijs.d.ts" />
+/// <reference path="../../definitions/lib/Q.d.ts" />
 /// <reference path="../../ts-mortar/promises/tspromises.d.ts" />
 /// <reference path="../../ts-mortar/events/events.d.ts" />
 /// <reference path="../../ts-code-generator/code-types/ast-types.d.ts" />