+++ change-trackers/collection-changes.d.ts
@@ -4,14 +4,24 @@
  */
 declare module Changes {
 
+    /** The API for a collection change tracker */
     export interface CollectionChangeTracker {
+
+        /** Inform this change tracker that the following items have been added */
         addChangeItemsAdded(items: any | any[]);
+
+        /** Inform this change tracker that the following items have been modified */
         addChangeItemsModified(items: any | any[]);
+
+        /** Inform this change tracker that the following items have been removed */
         addChangeItemsRemoved(items: any | any[]);
+
+        /** Inform this change tracker that the following change event has occured */
         addChange(change: CollectionChange);
     }
 
 
+    /** A collection change event */
     export interface CollectionChange {
         originalCount?: number;
         added: number;